Refactor toolAgent.ts into modular components
