+++ .changeset/refactor-tool-agent.md
@@ -1,0 +1,5 @@
+---
+"mycoder-agent": minor
+---
+
+Refactored toolAgent.ts into modular components for improved maintainability and testability. Split into config.ts, messageUtils.ts, toolExecutor.ts, tokenTracking.ts, and types.ts modules.
